+++ cmd/gwif/gcloudExec.go
@@ -8,19 +8,18 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var execCmd = &cobra.Command{
-	Use:   "exec -- [gcloud commands and arguments]",
+var gcloudExec = &cobra.Command{
+	Use:   "gcloud-exec -- [gcloud commands and arguments]",
 	Short: "Execute a gcloud command with WIF authentication",
 	Long: `Execute a gcloud command with Google Cloud Workload Identity Federation authentication.
 The -- separator MUST be used to separate sparkci command from gcloud arguments.
 
 Example:
-  sparkci gwif exec -- secrets versions access latest --project="my-project"`,
+  sparkci gwif gcloud-exec -- secrets versions access latest --project="my-project"`,
 	// Don't validate args so we can handle them ourselves
 	DisableFlagParsing: true,
 	SilenceErrors:      true,
 	SilenceUsage:       true,
-	SuggestFor:         []string{"exec", "gwif"},
 	RunE: func(cmd *cobra.Command, args []string) error {
 		// Look for the -- separator
 		var gcloudArgs []string
