+++ cmd/gwif/printSaToken.go
